Rewrite Lapse implementation from the top, again
Was hoping to figure out which part is causing the PS4 to panic. Not
much luck with that; it's the initial aio_multi_delete race but I just
don't know how to fix it, or if it even is fixable :(

Did fix various issues while I was at it, notably some unfortunate
misuse of getsockopt. Somehow it returned valid addresses anyway? o_.

Also added an experimental SINGLECORE mode that copies what lapse.mjs
and lapse.lua do, but it doesn't actually seem to work, so.